fix(staging): transactional snapshot version, cached-version seed, deterministic page tests
