+++ client/src/pages/Staging.test.jsx
@@ -1,7 +1,9 @@
-import { describe, it, expect, vi, beforeEach } from 'vitest';
-import { render, screen, waitFor, fireEvent } from '@testing-library/react';
+import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
+import { render, screen, waitFor, fireEvent, cleanup } from '@testing-library/react';
 import { MemoryRouter } from 'react-router-dom';
 import apiClient from '../utils/api';
+import stagingCache from '../utils/stagingCache';
+import Staging from './Staging';
 
 vi.mock('../components/AccessControl', () => ({
   default: ({ children }) => children,
@@ -54,54 +56,53 @@ function candidate(id, firstName) {
   };
 }
 
-// Each snapshot carries the server's monotonic stamp for the candidate read.
+// The page reads one endpoint: every resource below comes from a single database
+// transaction stamped with `snapshotVersion`.
 function snapshot({ candidates, snapshotVersion }) {
   return (endpoint) => {
-    switch (endpoint) {
-      case '/admin/staging/candidates':
-        return Promise.resolve({ candidates, snapshotVersion });
-      case '/admin/cycles/active':
-        return Promise.resolve({ id: 'cycle-1', name: 'Fall 2026' });
-      case '/admin/applications':
-        return Promise.resolve({ applications: [] });
-      case '/admin/events':
-        return Promise.resolve([]);
-      case '/admin/review-teams':
-        return Promise.resolve([]);
-      case '/admin/existing-decisions':
-        return Promise.resolve({ perRoundDecisions: { resume: {}, coffee: {}, firstRound: {}, final: {} } });
-      default:
-        return Promise.resolve([]);
+    if (endpoint === '/admin/staging/snapshot') {
+      return Promise.resolve({
+        snapshotVersion,
+        candidates,
+        activeCycle: { id: 'cycle-1', name: 'Fall 2026' },
+        applications: [],
+        events: [],
+        reviewTeams: [],
+        perRoundDecisions: { resume: {}, coffee: {}, firstRound: {}, final: {} },
+      });
     }
+    return Promise.resolve([]);
   };
 }
 
 async function renderStaging() {
-  // The page keeps a module-level snapshot cache, so reload it per test.
-  vi.resetModules();
-  const { default: Staging } = await import('./Staging');
   render(
     <MemoryRouter>
       <Staging />
     </MemoryRouter>
   );
-  await screen.findByText('Alice Example');
 }
 
 describe('Staging sync', () => {
   beforeEach(() => {
     vi.clearAllMocks();
+    // The first-paint cache outlives a render, so every test starts from an empty one.
+    stagingCache.invalidate();
     apiClient.get.mockImplementation(snapshot({ candidates: [candidate('c1', 'Alice')], snapshotVersion: 200 }));
+  });
+
+  afterEach(() => {
+    cleanup();
+    stagingCache.invalidate();
   });
 
   it('keeps the last successful snapshot on screen when a poll fails', async () => {
     await renderStaging();
+    await screen.findByText('Alice Example');
     expect(screen.getByText(/^Synced /)).toBeInTheDocument();
 
-    apiClient.get.mockImplementation((endpoint) =>
-      endpoint === '/admin/staging/candidates'
-        ? Promise.reject(new Error('Failed to load staging candidates'))
-        : snapshot({ candidates: [], snapshotVersion: 300 })(endpoint)
+    apiClient.get.mockImplementation(() =>
+      Promise.reject(new Error('Failed to load staging snapshot'))
     );
 
     fireEvent.click(screen.getByLabelText('Refresh staging data'));
@@ -110,8 +111,9 @@ describe('Staging sync', () => {
     expect(screen.getByText('Alice Example')).toBeInTheDocument();
   });
 
-  it('rejects a snapshot the server read before the one already applied', async () => {
+  it('rejects a snapshot the database read before the one already applied', async () => {
     await renderStaging();
+    await screen.findByText('Alice Example');
 
     apiClient.get.mockImplementation(
       snapshot({ candidates: [candidate('c2', 'Bob')], snapshotVersion: 100 })
@@ -119,15 +121,14 @@ describe('Staging sync', () => {
 
     fireEvent.click(screen.getByLabelText('Refresh staging data'));
 
-    await waitFor(() =>
-      expect(apiClient.get).toHaveBeenCalledWith('/admin/staging/candidates', expect.anything())
-    );
+    await waitFor(() => expect(apiClient.get).toHaveBeenCalledTimes(2));
     await waitFor(() => expect(screen.queryByText('Bob Example')).not.toBeInTheDocument());
     expect(screen.getByText('Alice Example')).toBeInTheDocument();
   });
 
-  it('applies a snapshot the server read after the one already applied', async () => {
+  it('applies a snapshot the database read after the one already applied', async () => {
     await renderStaging();
+    await screen.findByText('Alice Example');
 
     apiClient.get.mockImplementation(
       snapshot({ candidates: [candidate('c2', 'Bob')], snapshotVersion: 300 })
@@ -137,5 +138,30 @@ describe('Staging sync', () => {
 
     await screen.findByText('Bob Example');
     expect(screen.queryByText('Alice Example')).not.toBeInTheDocument();
+  });
+
+  it('rejects a snapshot older than the cached one it remounted with', async () => {
+    apiClient.get.mockImplementation(
+      snapshot({ candidates: [candidate('c1', 'Alice')], snapshotVersion: 300 })
+    );
+    await renderStaging();
+    await screen.findByText('Alice Example');
+
+    // Leave and come back: the page paints from the cache, so version 300 is on screen
+    // even though this poller has fetched nothing yet.
+    cleanup();
+    apiClient.get.mockImplementation(
+      snapshot({ candidates: [candidate('c2', 'Bob')], snapshotVersion: 100 })
+    );
+    await renderStaging();
+    await screen.findByText('Alice Example');
+
+    fireEvent.click(screen.getByLabelText('Refresh staging data'));
+
+    await waitFor(() =>
+      expect(apiClient.get).toHaveBeenCalledWith('/admin/staging/snapshot', expect.anything())
+    );
+    await waitFor(() => expect(screen.queryByText('Bob Example')).not.toBeInTheDocument());
+    expect(screen.getByText('Alice Example')).toBeInTheDocument();
   });
 });
